Page Admin Dashboard dan memberikan akses fitur fitur gudang

+++ database/migrations/2025_12_13_063511_create_services_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('services', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('price', 15, 2);
+            $table->bigInteger('price');
             $table->timestamps();
         });
     }
